requirements.txt
anyjson==0.3.3
billiard==3.3.0.19
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.1
django-shortuuidfield==0.1.2
django-toolbelt==0.0.1
feedparser==5.1.3
gunicorn==19.1.1
psycopg2==2.5.4
pytz==2014.10
requests==0.13.2
shortuuid==0.4.2
static3==0.5.1
stripe==1.19.1
